code reconstrcut
