+++ tianji/knowledges/RQA/model/__init__.py
@@ -1,0 +1,1 @@
+from .langchain.model import InternLM_LLM, Zhipu_LLM
