--- tianji/knowledges/script/__init__.py
@@ -1,1 +1,0 @@
-from .config import RQA_ST_Liyi_Chroma_Config
