--- tianji/knowledges/script/model/__init__.py
@@ -1,2 +1,0 @@
-from .internlm_llm import InternLM_LLM
-from .zhipu_llm import Zhipu_LLM
